endian.h: fix runtime breakage on LP64 platforms
LP64 platforms use a 64-bit integer for longs and pointers. Expecting 'unsigned
long' to be 32-bit in this case totally breaks.

+++ endian.h
@@ -23,8 +23,8 @@ typedef unsigned char uint8;
 typedef signed char int8;
 typedef unsigned short uint16;
 typedef signed short int16;
-typedef unsigned long uint32;
-typedef signed long int32;
+typedef unsigned int uint32;
+typedef signed int int32;
 
 //FCS added for windows build
 #ifdef _WIN32
